Rename to `property-cached` to release fork
