+++ LICENSE
@@ -1,4 +1,5 @@
-Copyright (c) 2015, Daniel Greenfeld
+Copyright (c) 2018-2019, Martin Larralde
+Copyright (c) 2015-2018, Daniel Greenfeld
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
