+++ tests/test_coroutine_cached_property.py
@@ -9,7 +9,7 @@ import unittest
 import asyncio
 from freezegun import freeze_time
 
-import cached_property
+import property_cached as  cached_property
 
 
 def unittest_run_loop(f):
